requirements.txt
streamlit
pymongo==4.6.2
google-genai==0.3.0
tenacity==8.2.3
deep-translator==1.11.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
langdetect==1.0.9
reportlab==4.1.0
streamlit-aggrid
pandas
plotly
matplotlib